backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
bcrypt
google-genai
pypdf
python-docx
beautifulsoup4
httpx
speechrecognition
pyannote.audio
google-cloud-vision